be/requirements.txt
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.6.0
Flask-SQLAlchemy==3.1.1
Flask-Cors==4.0.1
sendgrid==6.11.0
python-dotenv==1.0.1
pymysql==1.1.1
selenium==4.23.0
pytest-flask==1.3.0
webdriver-manager==4.0.1
flask-socketio==5.3.6
chromedriver-py==127.0.6533.72


